removed main scene cab and source files

+++ fastfingers.spritebuilder/Source/Gameplay.m
@@ -420,7 +420,7 @@ typedef NS_ENUM(NSInteger, DrawingOrder) {
 
 -(void)shareImage{
    [clickSound play];
-   NSString *message = [NSString stringWithFormat:@"Hey!!! I scored %d", _points];
+   NSString *message = [NSString stringWithFormat:@"Hey!!! I scored %ld", (long)_points];
    message = [message stringByAppendingString:@" points in Fast Fingers."];
    
    UIActivityViewController *activityVC = [[UIActivityViewController alloc] initWithActivityItems:[NSArray arrayWithObjects:message,_image, nil] applicationActivities:nil];
